requirements.txt
Flask
psycopg2-binary
PyJWT
cryptography
black
isort
pylint
